Add files via upload
Gears in a Robot: 
As one of the main mechanical elements (transmission components), gears are used in robots, for example, as a component of speed reducers inside manipulators. There, the rotation of a servomotor (electric actuator) is slowed down using a speed reducer to amplify its output torque, making it possible to obtain a large force with a small motor. In the case of robots, planetary gear mechanisms, strain wave gearing, and cycloid reduction gears are employed as speed reducer types, and are used according to the application. Gears are also indispensable and important mechanical elements used in many other places besides speed reducers, where they are required to increase or decrease speed, amplify torque, change the direction of rotation, and so on. Gears used in robots are processed as needed, such as tooth surface grinding when high precision and quietness are required, and heat treatment when high strength is required.

Reference link: https://www.seimitsu.in/Blog-on-Gears-For-Robotics.html